optimized the utils files

+++ features/media/mediapicker/src/main/java/com/tokopedia/media/picker/ui/component/CameraViewComponent.kt
@@ -95,7 +95,7 @@ class CameraViewComponent(
 
         cameraView.takeVideoSnapshot(
             FileCamera.createVideo(),
-            param.maxVideoDuration().toInt()
+            param.maxVideoDuration()
         )
     }
